+++ src/components/InkEditor/components/BlockHandles/BlockHandles.types.ts
@@ -1,0 +1,6 @@
+export interface BlockHandlesProps {
+  visible: boolean;
+  top: number;
+  onMoveUp: () => void;
+  onMoveDown: () => void;
+}